// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.4"
  }
  public: {
    Tables: {
// ...
      rooms: {
        Row: {
          capacity: number
          created_at: string | null
          description: string | null
          ical_export_url: string | null
          ical_import_url: string | null
          id: string
          image_urls: string[] | null
          name: string
          price: number
          status: string | null
        }
        Insert: {
          capacity: number
          created_at?: string | null
          description?: string | null
          ical_export_url?: string | null
          ical_import_url?: string | null
          id?: string
          image_urls?: string[] | null
          name: string
          price: number
          status?: string | null
        }
        Update: {
          capacity?: number
          created_at?: string | null
          description?: string | null
          ical_export_url?: string | null
          ical_import_url?: string | null
          id?: string
          image_urls?: string[] | null
          name?: string
          price?: number
          status?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}